# Optimize eval syms