This PR extends the `NamedConstants` in the `Store` to also hold `Ptr`s — and uses those during evaluation rather than constantly calling `Store::lurk_sym(…)`. These calls made evaluation become horribly slow when hierarchical symbols were introduced, because of a very bad and unperformant initial implementation. That performance should be improved when some of the LDON work lands. However, even if symbol interning becomes 'cheap', it's still an expensive operation to perform many times per reduction. This PR therefore fixes the root performance problem, which predated the symbol change.

See benchmark results below for an 11x speedup (-91% time).

```
Running benches/eval.rs (target/release/deps/eval-60f9add2215a78cb)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

Gnuplot not found, using plotters backend
go_base_10_16_bls12     time:   [6.9912 ms 6.9979 ms 7.0048 ms]                                
                        change: [-91.434% -91.420% -91.407%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high mild

Benchmarking go_base_10_160_bls12: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 6.7s, or reduce sample count to 70.
go_base_10_160_bls12    time:   [66.837 ms 66.909 ms 66.995 ms]                                 
                        change: [-91.236% -91.215% -91.195%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high severe

go_base_10_16_pasta_pallas                                                                            
                        time:   [6.8763 ms 6.8836 ms 6.8912 ms]
                        change: [-91.304% -91.289% -91.274%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

Benchmarking go_base_10_160_pasta_pallas: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 6.6s, or reduce sample count to 70.
go_base_10_160_pasta_pallas                                                                            
                        time:   [65.422 ms 65.482 ms 65.545 ms]
                        change: [-91.190% -91.176% -91.162%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
```

I then ran the benchmark on aea8116c, the commit preceding #180 (which introduced hierarchical symbols). As you can see, the older commit was 68% slower, so this is a strict improvement.


```
     Running benches/eval.rs (target/release/deps/eval-7dda34dad2478af3)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

Gnuplot not found, using plotters backend
go_base_10_16_bls12     time:   [11.389 ms 11.419 ms 11.457 ms]                                
                        change: [+62.724% +63.181% +63.802%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe

Benchmarking go_base_10_160_bls12: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 10.7s, or reduce sample count to 40.
go_base_10_160_bls12    time:   [106.88 ms 107.19 ms 107.57 ms]                                 
                        change: [+59.707% +60.198% +60.759%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) high mild
  5 (5.00%) high severe

go_base_10_16_pasta_pallas                                                                            
                        time:   [11.781 ms 11.794 ms 11.808 ms]
                        change: [+71.079% +71.339% +71.597%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) low mild
  6 (6.00%) high mild

Benchmarking go_base_10_160_pasta_pallas: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 11.1s, or reduce sample count to 40.
go_base_10_160_pasta_pallas                                                                            
                        time:   [110.15 ms 110.30 ms 110.45 ms]
                        change: [+68.163% +68.436% +68.715%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
```

For reference, the `go_base_10_160_pasta_pallas` test performs 10,618 iterations — so current performance (this PR) is ~163K iterations/second.